# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ludus UX API
  description: |
    REST API for the Ludus UX web application — a Next.js frontend for managing
    [Ludus](https://ludus.cloud) cyber range infrastructure, including VM ranges,
    content sources (sync / force re-install), templates, user management,
    VNC/SPICE console access, GOAD (Game of Active Directory) lab orchestration,
    and LudusHound BloodHound→Ludus deploys.

    ## Authentication

    All protected endpoints require a valid session cookie set by `POST /api/auth/login`.
    The cookie is AES-256-GCM encrypted and HTTP-only. Sessions expire after 24 hours.

    **Admin impersonation headers** (admin sessions only):
    - `X-Impersonate-As` — target username
    - `X-Impersonate-Apikey` — target user's Ludus API key
    - `X-Ludus-Admin: true` — route to admin API port (8081)
    - `X-Ludus-User` — override Ludus user context
  # Keep in sync with package.json / APP_VERSION (src/lib/changelog.ts).
  version: 1.4.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Ludus UX application server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ludus_session
      description: AES-256-GCM encrypted HTTP-only session cookie set by the login endpoint.
...
  schemas:
...
    TemplateSource:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        files:
          type: array
          items:
            type: string
        ref:
          type: string
        apiBase:
          type: string
          format: uri

    # ---- Sources (Ludus ≥2.2.0) ----
    LudusSource:
      type: object
      properties:
        id:
          type: string
        sourceID:
          type: string
          description: Canonical Ludus source identifier (preferred over `id`).
        name:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
        ref:
          type: string
          description: Tracked git branch, tag, or commit.
        lastSyncedAt:
          type: string
        lastSyncStatus:
          type: string
        lastSyncError:
          type: string
      additionalProperties: true
    SourceCatalogInstallState:
      type: string
      description: |
        Ludus catalog install hint. LUX treats `upgrade_available` (or a catalog
        vs installed version mismatch) as outdated content that can be force-reinstalled.
      enum: [not_installed, installed, upgrade_available]
...
    SourceCatalogItem:
      type: object
      properties:
        name:
          type: string
        fqcn:
          type: string
        version:
          type: string
        state:
          $ref: '#/components/schemas/SourceCatalogInstallState'
        scope:
          type: string
          description: e.g. `local` or `subscription` for Ansible items.
        description:
          type: string
        sourceBlueprintID:
          type: string
        min_ludus_version:
          type: string
      additionalProperties: true

    # ---- GOAD ----
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
  responses:
    Unauthorized:
      description: No valid session cookie or session has expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
...
paths:
...
  /api/templates/sources:
    get:
      tags: [Templates]
      summary: List available VM templates
      description: |
        Lists available VM templates from a remote GitLab repository
        (default: `badsectorlabs/ludus`). Results are cached in-process for
        5 minutes. No authentication required.
      operationId: listTemplateSources
      parameters:
        - name: source
          in: query
          required: false
          schema:
            type: string
            default: badsectorlabs
        - name: repoUrl
          in: query
          required: false
          schema:
            type: string
            format: uri
          description: Custom GitLab API base URL.
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: templates
          description: Path within the repository.
        - name: ref
          in: query
          required: false
          schema:
            type: string
            default: main
          description: Git ref (branch, tag, or commit SHA).
      responses:
        '200':
          description: Available templates.
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/TemplateSource'
...
  /api/sources:
    get:
      tags: [Sources]
      summary: List registered Ludus sources
      description: |
        Proxies Ludus `GET /sources`. When the connected Ludus lacks the Sources
        API (pre-2.2.0), returns `{ sources: [], available: false }` instead of an error.
        Also kicks LUX source auto-sync (stale git pulls) in the background when enabled.
      operationId: listSources
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Registered sources.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/LudusSource'
                  available:
                    type: boolean
                    description: "false when Ludus has no Sources API."
                  autoSyncIntervalMs:
                    type: integer
                    description: |
                      LUX source auto-sync interval / staleness threshold in milliseconds
                      (`SOURCE_AUTO_SYNC_INTERVAL_MS`, default 300000).
                    example: 300000
                required: [sources, available]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          description: Upstream Ludus error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Sources]
      summary: Register a git source
      description: |
        Registers a git repository as a Ludus source (`ludus source add`).
        Passes a distinct Ludus `id` (defaults to org-repo[-ref]) so the same
        URL can be registered once per branch/tag. If URL+ref already exists,
        returns that sourceID without creating a duplicate.
        Does not install catalog items; use install or the Add-from-Source UIs.
      operationId: registerSource
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  description: HTTPS or SSH git URL.
                ref:
                  type: string
                  default: main
                  description: Branch, tag, or commit to track.
                id:
                  type: string
                  description: |
                    Optional Ludus source id. Defaults to a slug of org/repo,
                    with `-<ref>` appended when ref is not main/master.
      responses:
        '200':
          description: Source registered (or already present for URL+ref).
          content:
            application/json:
              schema:
                type: object
                properties:
                  sourceID:
                    type: string
                  message:
                    type: string
        '400':
          description: Missing/invalid URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Ludus Sources API unavailable (need ≥2.2.0).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream Ludus error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sources/{sourceId}:
    patch:
      tags: [Sources]
      summary: Update a registered source
      description: |
        Change source metadata. When only `ref` is set, LUX deletes the source
        (no purge) and re-registers it for the new branch/tag — Ludus clones are
        often single-branch, so a bare PATCH+sync fails with git pathspec errors.
        Response may include a new `sourceID` and `recreated: true`.
      operationId: updateSource
      security:
        - sessionCookie: []
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ref:
                  type: string
                  description: New branch, tag, or commit.
                url:
                  type: string
                  format: uri
                  description: New remote URL for git-type sources.
      responses:
        '200':
          description: Source updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
                  sourceID:
                    type: string
                  ref:
                    type: string
        '400':
          description: Missing ref/url.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Ludus Sources API unavailable or source missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream Ludus error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Sources]
      summary: Delete a registered source
      description: |
        Removes the source registration. Installed templates/roles/collections
        remain on disk unless `purge` is true (Ludus purge semantics).
      operationId: deleteSource
      security:
        - sessionCookie: []
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                purge:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Source deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Ludus Sources API unavailable or source missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream Ludus error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/sources/{sourceId}/blueprints:
    get:
      tags: [Sources]
      summary: List blueprints in a source catalog
      operationId: listSourceBlueprints
      security:
        - sessionCookie: []
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Catalog blueprints.
          content:
            application/json:
              schema:
                type: object
                properties:
                  blueprints:
                    type: array
                    items:
                      $ref: '#/components/schemas/SourceCatalogItem'
                  catalogSource:
                    type: string
                    description: e.g. `ludus` or `github` fallback listing.
                  catalogRef:
                    type: string
                    description: Git ref from the Sources tab registration used for tip enrichment.
                  pins:
                    type: object
                    additionalProperties:
                      type: string
                    description: LUX install pins (name → version) from last successful install/re-sync.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Ludus Sources API unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...